extract control mode
