+++ engine/control_mode.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+class ControlMode(Enum):
+    """
+    An enumeration of different control modes
+    """
+    LAWNMOWER = 1
+    LAWNMOWER_B = 2
+    LAWNMOWER_A = 3
+    SPIRAL = 4
+    ROOMBA = 5
+    MANUAL = 6
+    STRAIGHT = 7
