+++ tests/functionality_tests/obstacle_tracking_test.py
@@ -4,7 +4,7 @@ from engine.robot import Robot
 from engine.robot_state import Robot_State
 from engine.mission import Mission
 from engine.base_station import BaseStation
-from engine.mission import ControlMode
+from engine.control_mode import ControlMode
 from constants.definitions import ROOT_DIR
 import random
